Finalize matrix

+++ ci/stop.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-docker kill $(docker ps -q)
-docker rm $(docker ps -a -q)
-docker rmi $(docker images -q)
+if [ "$MATRIX_JOB_TYPE" == "TEST" ]; then
+    docker kill $(docker ps -q)
+    docker rm $(docker ps -a -q)
+    docker rmi $(docker images -q)
+fi
